Add files via upload
A* Routing Program to find the shortest path between a starting node (which is either a node with fuel or a solar powered node) and a goal node from a grid input. On the grid there could be refuel locations or obstacles. We then print the grid with the shortest path, and the other paths that have been explored. The heuristic for A* is using the Manhattan distance.